Updated salary calculator class
Changed parameter type for a method

+++ TechChallenge/Assets/Scripts/Tools/Calculators/SalaryCalculator.cs
@@ -6,9 +6,9 @@ namespace Tools.Calculators
 {
     public static class SalaryCalculator 
     {
-        public static int CalculateSalaryIncrease(float increasePercentage, int currentSalary)
+        public static int CalculateSalaryIncrease(float increasePercentage, float currentSalary)
         {
-            float unroundedSalary = (float)currentSalary * (1 + (increasePercentage / 100));
+            float unroundedSalary = currentSalary * (1 + (increasePercentage / 100));
             return Mathf.RoundToInt(unroundedSalary);
         }
     }
